+++ tests/setupPoseidon.ts
@@ -1,0 +1,3 @@
+import { ensurePoseidonLoaded } from "../src";
+
+await ensurePoseidonLoaded();